UPDATE: change number type from int to double

+++ FileKnifeTool/src/logprocessing/AggregatingStatistic.java
@@ -11,7 +11,7 @@ import org.slf4j.LoggerFactory;
 public class AggregatingStatistic extends StatisticDefinition {
 
 	
-		private Integer counter=0;
+		private Double counter=0.0;
 		//private String regexp="";
 		private Map stats=new HashMap<String, HashMap>();
 		//private String statname="";
